Split Identity into some submodules
This commit pulls the Login and Alias concepts out of the
`ledger/identity.js` file. The rationale is that as part of #2109 we
will need to make multiple versions of the Identity type (and, for
backcompat reasons, support all versions indefinitely). Therefore the
file needs to be split lest it become a huge mess. This is the first and
easiest bit to pull out.

For now, I'm re-exporting the types, functions, and parsers from the
same `identity.js` file, so we don't need to change any clients.

Test plan: Simple re-org, `yarn test` passes.
